// src/interfaces/ITasks.ts
// Generated by https://quicktype.io

export interface IResTasks {
    data: Data;
}

export interface Data {
    tasksList: TasksList;
}

export interface TasksList {
    items: ITask[];
}

export interface ITask {
    id:          string;
    no:          number;
    name:        string;
    description: string;
    state:       TaskState;
    status:      number;
}

export enum TaskState {
    Pending = 'Pending',
    Processing = 'Processing',
    Completed = 'Completed'
}